% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec1east_bbox.R
\docType{data}
\name{sec1east_bbox}
\alias{sec1east_bbox}
\title{sec1east_bbox}
\format{
A data frame with 1 rows and 2 variables:
\describe{
\item{\code{id}}{double ID of boundary.}
\item{\code{geometry}}{list Coordinates of bounding box polygon.}
}
}
\usage{
sec1east_bbox
}
\description{
Field boundary for Farmer B's OFPE experimental field
named 'sec1east'.
}
\details{
DETAILS
}
\keyword{datasets}
