% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipShapefiles.R
\name{unzipShapefiles}
\alias{unzipShapefiles}
\title{unzipShapefiles}
\usage{
unzipShapefiles(zip_file, path)
}
\arguments{
\item{zip_file}{Character, the name of the zipped file to
extract.}

\item{path}{Character, path to the folder with the data.}
}
\value{
A shapefile of class 'sf'.
}
\description{
Unzips compressed files containing ESRI shapefiles from the
'data-raw' folder of the OFPEDATA package. Takes a zipped file
name and path to data and checks if there are any shapefiles
in the zipped object and if so, returns a 'sf' object.
}
\examples{
\dontrun{
unzipShapefiles("zipped_data_filename", "path/to/data")
}

}
